+++ 2D_Game/Assets/Scripts/IvyInteract.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bddf9d72742d64195b2bfa338c3544cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
